# docs/API.openapi.yaml
openapi: 3.0.0
info:
  title: BuildBrainOS API
  description: |
    Complete REST API specification for BuildBrainOS platform.
    All endpoints require JWT authentication in Authorization header.
    
    Base URLs:
    - Development: http://localhost:4000/api/v1
    - Production: https://api.buildbrainos.com/api/v1
    
    Authentication:
    - JWT Bearer Token in Authorization header
    - Refresh tokens valid for 30 days
    - Rate limiting: 1000 requests/hour per user
  version: 1.0.0
  contact:
    name: BuildBrainOS Support
    email: support@buildbrainos.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: https://api.buildbrainos.com/api/v1
    description: Production server
...
paths:
...
  /users/profile:
    get:
      tags:
        - Users
      summary: Get current user profile
      operationId: getUserProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags:
        - Users
      summary: Update user profile
      operationId: updateUserProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                bio:
                  type: string
                avatar:
                  type: string
                  format: uri
                address:
                  $ref: '#/components/schemas/Address'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Get user profile by ID
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [GC, SUBCONTRACTOR, WORKER, GOVERNMENT, SUPPLIER]
        companyName:
          type: string
        kycTier:
          type: string
          enum: [TIER_1, TIER_2, TIER_3]
        kycStatus:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        profile:
          type: object
          properties:
            phone:
              type: string
            bio:
              type: string
            avatar:
              type: string
              format: uri
            address:
              $ref: '#/components/schemas/Address'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
        latitude:
          type: number
        longitude:
          type: number
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header
...